==> code/01-feature_engineering.ipynb <==
import numpy as np
import pandas as pd
import os, sys, gc, time, warnings, pickle, psutil, random

from math import ceil

from sklearn.preprocessing import LabelEncoder
warnings.filterwarnings('ignore')

# TODO

## 引入收入
https://www.kaggle.com/malavika8/walmartproject
df['revenue'] = df['sold']*df['sell_price'].astype(np.float32)


## [Downcasting](https://www.kaggle.com/anshuls235/m5-forecasting-eda-fe-modelling?select=submission.csv)

In this section I'll be downcasting the dataframes to reduce the amount of storage used by them and also to expidite the operations performed on them.

- **Numerical Columns**: Depending on your environment, pandas automatically creates int32, int64, float32 or float64 columns for numeric ones. If you know the min or max value of a column, you can use a subtype which is less memory consuming. You can also use an unsigned subtype if there is no negative value.
Here are the different subtypes you can use:
    - int8 / uint8 : consumes 1 byte of memory, range between -128/127 or 0/255
    - bool : consumes 1 byte, true or false
    - float16 / int16 / uint16: consumes 2 bytes of memory, range between -32768 and 32767 or 0/65535
    - float32 / int32 / uint32 : consumes 4 bytes of memory, range between -2147483648 and 2147483647
    - float64 / int64 / uint64: consumes 8 bytes of memory
    
If one of your column has values between 1 and 10 for example, you will reduce the size of that column from 8 bytes per row to 1 byte, which is more than 85% memory saving on that column!


- **Categorical Columns**: Pandas stores categorical columns as **objects**. One of the reason this storage is not optimal is that it creates a list of pointers to the memory address of each value of your column. For columns with low cardinality (the amount of unique values is lower than 50% of the count of these values), this can be optimized by forcing pandas to use a virtual mapping table where all unique values are mapped via an integer instead of a pointer. This is done using the category datatype.

def get_memory_usage():
    return np.round(psutil.Process(os.getpid()).memory_info()[0] / 2.**30, 2)


def sizeof_fmt(num, suffix='B'):
    for unit in ['', 'Ki', 'Mi', 'Gi', 'Ti', 'Pi', 'Ei', 'Zi']:
        if abs(num) < 1024.0:
            return "%3.1f%s%s" % (num, unit, suffix)
        num /= 1024.0
    return "%.1f%s%s" % (num, 'Yi', suffix)

def reduce_mem_usage(df, verbose=True):
    numerics = ['int16', 'int32', 'int64', 'float16', 'float32', 'float64']
    start_mem = df.memory_usage().sum() / 1024**2
    for col in df.columns:
        col_type = df[col].dtypes
        if col_type in numerics:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(
                        np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(
                        np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(
                        np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(
                        np.int64).max:
                    df[col] = df[col].astype(np.int64)
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(
                        np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(
                        np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)
    end_mem = df.memory_usage().sum() / 1024**2
    if verbose:
        print('Mem. usage decreased to {:5.2f} Mb ({:.1f}% reduction)'.format(
            end_mem, 100 * (start_mem - end_mem) / start_mem))
    return df

def merge_by_concat(df1, df2, merge_on):
    merged_gf = df1[merge_on]
    merged_gf = merged_gf.merge(df2, on=merge_on, how='left')
    new_columns = [col for col in list(merged_gf) if col not in merge_on]
    df1 = pd.concat([df1, merged_gf[new_columns]], axis=1)
    return df1

########################### Vars
#################################################################################
TARGET = 'sales'  # Our main target
END_TRAIN = 1941  # Last day in train set
MAIN_INDEX = ['id', 'd']  # We can identify item by these columns

########################### Load Data
#################################################################################
print('Load Main Data')

train_df = pd.read_csv('../data/sales_train_evaluation.csv')
prices_df = pd.read_csv('../data/sell_prices.csv')
calendar_df = pd.read_csv('../data/calendar.csv')

train_df.head()

train_df.shape

########################### Make Grid
#################################################################################
print('Create Grid')

index_columns = ['id','item_id','dept_id','cat_id','store_id','state_id']
grid_df = pd.melt(train_df, 
                  id_vars = index_columns, 
                  var_name = 'd', 
                  value_name = TARGET)


print('Train rows:', len(train_df), len(grid_df))
grid_df.head()

#### 数据截断

# grid_df['day'] = grid_df['d'].apply(lambda x: x[2:]).astype(np.int16)
# grid_df = grid_df[grid_df['day']>703]
# print(grid_df.shape) 
# # grid_df = grid_df.iloc[:1000, :]
# grid_df.head()

add_grid = pd.DataFrame()
for i in range(1,29):
    temp_df = train_df[index_columns]
    temp_df = temp_df.drop_duplicates()
    temp_df['d'] = 'd_'+ str(END_TRAIN+i)
    temp_df[TARGET] = np.nan
    add_grid = pd.concat([add_grid,temp_df])

grid_df = pd.concat([grid_df,add_grid])
grid_df = grid_df.reset_index(drop=True)

del temp_df, add_grid
del train_df
print("{:>20}: {:>8}".format('Original grid_df',sizeof_fmt(grid_df.memory_usage(index=True).sum())))

for col in index_columns:
    grid_df[col] = grid_df[col].astype('category')

print("{:>20}: {:>8}".format('Reduced grid_df',sizeof_fmt(grid_df.memory_usage(index=True).sum())))

grid_df.head()

########################### Product Release date
#################################################################################
print('Release week')


release_df = prices_df.groupby(['store_id','item_id'])['wm_yr_wk'].agg(['min']).reset_index()
release_df.columns = ['store_id','item_id','release']


grid_df = merge_by_concat(grid_df, release_df, ['store_id','item_id'])
del release_df

grid_df = merge_by_concat(grid_df, calendar_df[['wm_yr_wk','d']], ['d'])
                      
grid_df = grid_df[grid_df['wm_yr_wk']>=grid_df['release']]
grid_df = grid_df.reset_index(drop=True)

print("{:>20}: {:>8}".format('Original grid_df',sizeof_fmt(grid_df.memory_usage(index=True).sum())))

grid_df['release'] = grid_df['release'] - grid_df['release'].min()
grid_df['release'] = grid_df['release'].astype(np.int16)

print("{:>20}: {:>8}".format('Reduced grid_df',sizeof_fmt(grid_df.memory_usage(index=True).sum())))

########################### Save part 1
#################################################################################
print('Save Part 1')

grid_df.to_pickle('../data/output/grid_part_1.pkl')

print('Size:', grid_df.shape)

########################### Prices
#################################################################################
print('Prices')

prices_df['price_max'] = prices_df.groupby(['store_id', 'item_id'
                                            ])['sell_price'].transform('max')
prices_df['price_min'] = prices_df.groupby(['store_id', 'item_id'
                                            ])['sell_price'].transform('min')
prices_df['price_std'] = prices_df.groupby(['store_id', 'item_id'
                                            ])['sell_price'].transform('std')
prices_df['price_mean'] = prices_df.groupby(['store_id', 'item_id'
                                             ])['sell_price'].transform('mean')

prices_df['price_norm'] = prices_df['sell_price'] / prices_df['price_max']

prices_df['price_nunique'] = prices_df.groupby(
    ['store_id', 'item_id'])['sell_price'].transform('nunique')
prices_df['item_nunique'] = prices_df.groupby(
    ['store_id', 'sell_price'])['item_id'].transform('nunique')

calendar_prices = calendar_df[['wm_yr_wk', 'month', 'year']]
calendar_prices = calendar_prices.drop_duplicates(subset=['wm_yr_wk'])
prices_df = prices_df.merge(calendar_prices[['wm_yr_wk', 'month', 'year']],
                            on=['wm_yr_wk'],
                            how='left')
del calendar_prices

prices_df['price_momentum'] = prices_df['sell_price'] / prices_df.groupby(
    ['store_id', 'item_id'])['sell_price'].transform(lambda x: x.shift(1))
prices_df['price_momentum_m'] = prices_df['sell_price'] / prices_df.groupby(
    ['store_id', 'item_id', 'month'])['sell_price'].transform('mean')
prices_df['price_momentum_y'] = prices_df['sell_price'] / prices_df.groupby(
    ['store_id', 'item_id', 'year'])['sell_price'].transform('mean')

del prices_df['month'], prices_df['year']

########################### Merge prices and save part 2
#################################################################################
print('Merge prices and save part 2')

original_columns = list(grid_df)
grid_df = grid_df.merge(prices_df,
                        on=['store_id', 'item_id', 'wm_yr_wk'],
                        how='left')
keep_columns = [col for col in list(grid_df) if col not in original_columns]
grid_df = grid_df[MAIN_INDEX + keep_columns]
grid_df = reduce_mem_usage(grid_df)

# Safe part 2
grid_df.to_pickle('../data/output/grid_part_2.pkl')
print('Size:', grid_df.shape)

# We don't need prices_df anymore
del prices_df

# We can remove new columns
# or just load part_1
grid_df = pd.read_pickle('../data/output/grid_part_1.pkl')

grid_df = grid_df[MAIN_INDEX]

icols = [
    'date', 'd', 'event_name_1', 'event_type_1', 'event_name_2',
    'event_type_2', 'snap_CA', 'snap_TX', 'snap_WI'
]

grid_df = grid_df.merge(calendar_df[icols], on=['d'], how='left')

# Minify data
# 'snap_' columns we can convert to bool or int8
icols = [
    'event_name_1', 'event_type_1', 'event_name_2', 'event_type_2', 'snap_CA',
    'snap_TX', 'snap_WI'
]
for col in icols:
    grid_df[col] = grid_df[col].astype('category')

# Convert to DateTime
grid_df['date'] = pd.to_datetime(grid_df['date'])

# Make some features from date
grid_df['tm_d'] = grid_df['date'].dt.day.astype(np.int8)
grid_df['tm_w'] = grid_df['date'].dt.week.astype(np.int8)
grid_df['tm_m'] = grid_df['date'].dt.month.astype(np.int8)
grid_df['tm_y'] = grid_df['date'].dt.year
grid_df['tm_y'] = (grid_df['tm_y'] - grid_df['tm_y'].min()).astype(np.int8)
grid_df['tm_wm'] = grid_df['tm_d'].apply(lambda x: ceil(x / 7)).astype(np.int8)
grid_df['tm_dw'] = grid_df['date'].dt.dayofweek.astype(np.int8)
grid_df['tm_w_end'] = (grid_df['tm_dw'] >= 5).astype(np.int8)

# Remove date
del grid_df['date']

########################### Save part 3 (Dates)
#################################################################################
print('Save part 3')

grid_df.to_pickle('../data/output/grid_part_3.pkl')
print('Size:', grid_df.shape)

del calendar_df
del grid_df

## Part 1
# Convert 'd' to int
grid_df = pd.read_pickle('../data/output/grid_part_1.pkl')
grid_df['d'] = grid_df['d'].apply(lambda x: x[2:]).astype(np.int16)

# Remove 'wm_yr_wk'
# as test values are not in train set
del grid_df['wm_yr_wk']
grid_df.to_pickle('../data/output/grid_part_1.pkl')

del grid_df

# Now we have 3 sets of features
grid_df = pd.concat([pd.read_pickle('../data/output/grid_part_1.pkl'),
                     pd.read_pickle('../data/output/grid_part_2.pkl').iloc[:,2:],
                     pd.read_pickle('../data/output/grid_part_3.pkl').iloc[:,2:]],
                     axis=1)
                     
# Let's check again memory usage
print("{:>20}: {:>8}".format('Full Grid',sizeof_fmt(grid_df.memory_usage(index=True).sum())))
print('Size:', grid_df.shape)

# 2.5GiB + is is still too big to train our model
# (on kaggle with its memory limits)
# and we don't have lag features yet
# But what if we can train by state_id or shop_id?
state_id = 'CA'
grid_df = grid_df[grid_df['state_id']==state_id]
print("{:>20}: {:>8}".format('Full Grid',sizeof_fmt(grid_df.memory_usage(index=True).sum())))
#           Full Grid:   1.2GiB

store_id = 'CA_1'
grid_df = grid_df[grid_df['store_id']==store_id]
print("{:>20}: {:>8}".format('Full Grid',sizeof_fmt(grid_df.memory_usage(index=True).sum())))
#           Full Grid: 321.2MiB

# Seems its good enough now
# In other kernel we will talk about LAGS features
# Thank you.

########################### Final list of features
#################################################################################
grid_df.info()

grid_df.tail()
