==> code/04-groupkfold_final_lgbm.ipynb <==
# General imports
import numpy as np
import pandas as pd
import os, sys, gc, time, warnings, pickle, psutil, random
from sklearn.model_selection import GroupKFold
# custom imports
from multiprocessing import Pool  # Multiprocess Runs

warnings.filterwarnings('ignore')

- 全部采用tweedie_variance_power=1.25，score=0.48332
- STORES_IDS = ['CA_1', 'CA_2', 'CA_3', 'WI_2'] # tweedie_variance_power=1.05, 其他商店1.25。（score=0.47886）
- STORES_IDS = ['CA_1', 'CA_2', 'CA_3', 'WI_2'] # tweedie_variance_power=1.05, ['TX_1', 'TX_2', 'TX_3']1.1,其他商店1.25。（score=0.47566）
- STORES_IDS = ['CA_1', 'CA_2', 'CA_3', 'WI_2'] # tweedie_variance_power=1.05, ['TX_1', 'TX_2', 'TX_3', 'WI_3']1.1,['CA_4', 'WI_1']1.25。（score=0.47437）

-         if store_id in ['CA_3', 'WI_2']:
            lgb_params['tweedie_variance_power'] = 1.05
        elif store_id in ['WI_3']:
            lgb_params['tweedie_variance_power'] = 1.07
        elif store_id in ['CA_1', 'TX_1', 'TX_2']:
            lgb_params['tweedie_variance_power'] = 1.10
        elif store_id in ['CA_2', 'TX_3']:
            lgb_params['tweedie_variance_power'] = 1.15
        else:
            lgb_params['tweedie_variance_power'] = 1.25
（score=0.47523）

-         if store_id in ['CA_1', 'CA_2', 'CA_3', 'WI_2']:
            lgb_params['tweedie_variance_power'] = 1.05
        elif store_id in ['TX_1', 'TX_2', 'TX_3', 'WI_3']:
            lgb_params['tweedie_variance_power'] = 1.10
        else:
            lgb_params['tweedie_variance_power'] = 1.20
            
        if store_id in ['CA_3']:
            lgb_params['n_estimators'] = 1800
        else:
            lgb_params['n_estimators'] = 1400
（score=0.47421）


-         if store_id in ['CA_1', 'CA_2', 'CA_3', 'WI_2']:
            lgb_params['tweedie_variance_power'] = 1.05
        elif store_id in ['TX_1', 'TX_2', 'TX_3', 'WI_3']:
            lgb_params['tweedie_variance_power'] = 1.10
        else:
            lgb_params['tweedie_variance_power'] = 1.20
            
        if store_id in ['CA_2', 'CA_3']:
            lgb_params['n_estimators'] = 1800
        else:
            lgb_params['n_estimators'] = 1400
            
（score=0.47427）

########################### Helpers
#################################################################################
## Seeder
# :seed to make all processes deterministic     # type: int
def seed_everything(seed=0):
    random.seed(seed)
    np.random.seed(seed)


## Multiprocess Runs
def df_parallelize_run(func, t_split):
    num_cores = np.min([N_CORES, len(t_split)])
    pool = Pool(num_cores)
    df = pd.concat(pool.map(func, t_split), axis=1)
    pool.close()
    pool.join()
    return df

########################### Helper to load data by store ID
#################################################################################
# Read data
def get_data_by_store(store):

    # Read and contact basic feature
    df = pd.concat([
        pd.read_pickle(BASE),
        pd.read_pickle(PRICE).iloc[:, 2:],
        pd.read_pickle(CALENDAR).iloc[:, 2:]
    ],
                   axis=1)

    # Leave only relevant store
    df = df[df['store_id'] == store]

    # With memory limits we have to read
    # lags and mean encoding features
    # separately and drop items that we don't need.
    # As our Features Grids are aligned
    # we can use index to keep only necessary rows
    # Alignment is good for us as concat uses less memory than merge.
    df2 = pd.read_pickle(MEAN_ENC)[mean_features]
    df2 = df2[df2.index.isin(df.index)]

    df3 = pd.read_pickle(LAGS).iloc[:, 3:]
    df3 = df3[df3.index.isin(df.index)]

    df = pd.concat([df, df2], axis=1)
    del df2  # to not reach memory limit

    df = pd.concat([df, df3], axis=1)
    del df3  # to not reach memory limit

    # Create features list
    features = [col for col in list(df) if col not in remove_features]
    df = df[['id', 'd', TARGET] + features]

    # Skipping first n rows
    df = df[df['d'] >= START_TRAIN].reset_index(drop=True)

    return df, features


# Recombine Test set after training
def get_base_test():
    base_test = pd.DataFrame()

    for store_id in STORES_IDS:
        temp_df = pd.read_pickle(AUX_MODELS + 'test_' + store_id + '.pkl')
        temp_df['store_id'] = store_id
        base_test = pd.concat([base_test, temp_df]).reset_index(drop=True)

    return base_test


########################### Helper to make dynamic rolling lags
#################################################################################
def make_lag(LAG_DAY):
    lag_df = base_test[['id', 'd', TARGET]]
    col_name = 'sales_lag_' + str(LAG_DAY)
    lag_df[col_name] = lag_df.groupby([
        'id'
    ])[TARGET].transform(lambda x: x.shift(LAG_DAY)).astype(np.float16)
    return lag_df[[col_name]]


def make_lag_roll(LAG_DAY):
    shift_day = LAG_DAY[0]
    roll_wind = LAG_DAY[1]
    lag_df = base_test[['id', 'd', TARGET]]
    col_name = 'rolling_mean_tmp_' + str(shift_day) + '_' + str(roll_wind)
    lag_df[col_name] = lag_df.groupby(['id'])[TARGET].transform(
        lambda x: x.shift(shift_day).rolling(roll_wind).mean())
    return lag_df[[col_name]]

########################### Model params
#################################################################################
import lightgbm as lgb
lgb_params = {
    'boosting_type': 'gbdt',
    'objective': 'tweedie',
    'metric': 'rmse',
    'subsample': 0.5,
    'subsample_freq': 1,
    'learning_rate': 0.03,
    'num_leaves': 2**11 - 1,
    'min_data_in_leaf': 2**12 - 1,
    'feature_fraction': 0.5,
    'max_bin': 100,
    'n_estimators': 1400,
    'boost_from_average': False,
    'verbose': 1,
#     'device': 'gpu',
#     'gpu_platform_id': -1,
#     'gpu_device_id': -1
}

# Let's look closer on params

## 'boosting_type': 'gbdt'
# we have 'goss' option for faster training
# but it normally leads to underfit.
# Also there is good 'dart' mode
# but it takes forever to train
# and model performance depends
# a lot on random factor
# https://www.kaggle.com/c/home-credit-default-risk/discussion/60921

## 'objective': 'tweedie'
# Tweedie Gradient Boosting for Extremely
# Unbalanced Zero-inflated Data
# https://arxiv.org/pdf/1811.10192.pdf
# and many more articles about tweediie
#
# Strange (for me) but Tweedie is close in results
# to my own ugly loss.
# My advice here - make OWN LOSS function
# https://www.kaggle.com/c/m5-forecasting-accuracy/discussion/140564
# https://www.kaggle.com/c/m5-forecasting-accuracy/discussion/143070
# I think many of you already using it (after poisson kernel appeared)
# (kagglers are very good with "params" testing and tuning).
# Try to figure out why Tweedie works.
# probably it will show you new features options
# or data transformation (Target transformation?).

## 'tweedie_variance_power': 1.1
# default = 1.5
# set this closer to 2 to shift towards a Gamma distribution
# set this closer to 1 to shift towards a Poisson distribution
# my CV shows 1.1 is optimal
# but you can make your own choice

## 'metric': 'rmse'
# Doesn't mean anything to us
# as competition metric is different
# and we don't use early stoppings here.
# So rmse serves just for general
# model performance overview.
# Also we use "fake" evaluation set
# (as it makes part of the training set)
# so even general rmse score doesn't mean anything))
# https://www.kaggle.com/c/m5-forecasting-accuracy/discussion/133834

## 'subsample': 0.5
# Serves to fight with overfit
# this will randomly select part of data without resampling
# Chosen by CV (my CV can be wrong!)
# Next kernel will be about CV

##'subsample_freq': 1
# frequency for bagging
# default value - seems ok

## 'learning_rate': 0.03
# Chosen by CV
# Smaller - longer training
# but there is an option to stop
# in "local minimum"
# Bigger - faster training
# but there is a chance to
# not find "global minimum" minimum

## 'num_leaves': 2**11-1
## 'min_data_in_leaf': 2**12-1
# Force model to use more features
# We need it to reduce "recursive"
# error impact.
# Also it leads to overfit
# that's why we use small
# 'max_bin': 100

## l1, l2 regularizations
# https://towardsdatascience.com/l1-and-l2-regularization-methods-ce25e7fc831c
# Good tiny explanation
# l2 can work with bigger num_leaves
# but my CV doesn't show boost

## 'n_estimators': 1400
# CV shows that there should be
# different values for each state/store.
# Current value was chosen
# for general purpose.
# As we don't use any early stopings
# careful to not overfit Public LB.

##'feature_fraction': 0.5
# LightGBM will randomly select
# part of features on each iteration (tree).
# We have maaaany features
# and many of them are "duplicates"
# and many just "noise"
# good values here - 0.5-0.7 (by CV)

## 'boost_from_average': False
# There is some "problem"
# to code boost_from_average for
# custom loss
# 'True' makes training faster
# BUT carefull use it
# https://github.com/microsoft/LightGBM/issues/1514
# not our case but good to know cons

########################### Vars
#################################################################################
SEED = 42  # We want all things
seed_everything(SEED)  # to be as deterministic
lgb_params['seed'] = SEED  # as possible
N_CORES = psutil.cpu_count()  # Available CPU cores

#LIMITS and const
TARGET = 'sales'  # Our target
START_TRAIN = 0  # We can skip some rows (Nans/faster training)
END_TRAIN = 1941  # End day of our train set
P_HORIZON = 28  # Prediction horizon
USE_AUX = False  # Use or not pretrained models

#FEATURES to remove
## These features lead to overfit
## or values not present in test set
remove_features = [
    'id', 'state_id', 'store_id', 'date', 'wm_yr_wk', 'd', TARGET
]
mean_features = [
    'enc_cat_id_mean', 'enc_cat_id_std', 'enc_dept_id_mean', 'enc_dept_id_std',
    'enc_item_id_mean', 'enc_item_id_std'
]

#PATHS for Features
ORIGINAL = '../data/'
BASE = '../data/output/grid_part_1.pkl'
PRICE = '../data/output/grid_part_2.pkl'
CALENDAR = '../data/output/grid_part_3.pkl'
LAGS = '../data/output/lags_df_28.pkl'
MEAN_ENC = '../data/output/mean_encoding_df.pkl'

# AUX(pretrained) Models paths
AUX_MODELS = '../data/output/'

#STORES ids
STORES_IDS = pd.read_csv(ORIGINAL+'sales_train_evaluation.csv')['store_id']
STORES_IDS = list(STORES_IDS.unique())
# STORES_IDS = [
#     'CA_1'
# ]  # It takes around 5hrs to train for each store. Please run multiple kernels for each store.

#FOLDS
CV_FOLDS = [0, 1, 2]

#SPLITS for lags creation
SHIFT_DAY = 28
N_LAGS = 15
LAGS_SPLIT = [col for col in range(SHIFT_DAY, SHIFT_DAY + N_LAGS)]
ROLS_SPLIT = []
for i in [1, 7, 14]:
    for j in [7, 14, 30, 60]:
        ROLS_SPLIT.append([i, j])

########################### Aux Models
# If you don't want to wait hours and hours
# to have result you can train each store
# in separate kernel and then just join result.

# If we want to use pretrained models we can
## skip training
## (in our case do dummy training
##  to show that we are good with memory
##  and you can safely use this (all kernel) code)
# if USE_AUX:
#     lgb_params['n_estimators'] = 2

# Here is some 'logs' that can compare
#Train CA_1
#[100]	valid_0's rmse: 2.02289
#[200]	valid_0's rmse: 2.0017
#[300]	valid_0's rmse: 1.99239
#[400]	valid_0's rmse: 1.98471
#[500]	valid_0's rmse: 1.97923
#[600]	valid_0's rmse: 1.97284
#[700]	valid_0's rmse: 1.96763
#[800]	valid_0's rmse: 1.9624
#[900]	valid_0's rmse: 1.95673
#[1000]	valid_0's rmse: 1.95201
#[1100]	valid_0's rmse: 1.9476
#[1200]	valid_0's rmse: 1.9434
#[1300]	valid_0's rmse: 1.9392
#[1400]	valid_0's rmse: 1.93446

#Train CA_2
#[100]	valid_0's rmse: 1.88949
#[200]	valid_0's rmse: 1.84767
#[300]	valid_0's rmse: 1.83653
#[400]	valid_0's rmse: 1.82909
#[500]	valid_0's rmse: 1.82265
#[600]	valid_0's rmse: 1.81725
#[700]	valid_0's rmse: 1.81252
#[800]	valid_0's rmse: 1.80736
#[900]	valid_0's rmse: 1.80242
#[1000]	valid_0's rmse: 1.79821
#[1100]	valid_0's rmse: 1.794
#[1200]	valid_0's rmse: 1.78973
#[1300]	valid_0's rmse: 1.78552
#[1400]	valid_0's rmse: 1.78158

STORES_IDS

########################### Train Models
#################################################################################

for store_id in STORES_IDS:
    print('Train', store_id)

    # Get grid for current store
    grid_df, features_columns = get_data_by_store(store_id)

    # Masks for
    # Train (All data less than 1941)
    # "evaluation" (Last 28 days - not real validatio set)
    # Test (All data greater than 1941 day,
    #       with some gap for recursive features)
    train_mask = grid_df['d'] <= END_TRAIN
    preds_mask = grid_df['d'] > (END_TRAIN - 100)

    ## Initiating our GroupKFold
    folds = GroupKFold(n_splits=3)

    # get subgroups for each week, year pair
    grid_df['groups'] = grid_df['tm_w'].astype(
        str) + '_' + grid_df['tm_y'].astype(str)
    split_groups = grid_df[train_mask]['groups']

    # Main Data
    X, y = grid_df[train_mask][features_columns], grid_df[train_mask][TARGET]

    # Saving part of the dataset for later predictions
    # Removing features that we need to calculate recursively
    grid_df = grid_df[preds_mask].reset_index(drop=True)
    keep_cols = [col for col in list(grid_df) if '_tmp_' not in col]
    grid_df = grid_df[keep_cols]
    grid_df.to_pickle('../data/output/test_' + store_id + '.pkl')
    del grid_df

    # Launch seeder again to make lgb training 100% deterministic
    # with each "code line" np.random "evolves"
    # so we need (may want) to "reset" it

    for fold_, (trn_idx,
                val_idx) in enumerate(folds.split(X, y, groups=split_groups)):
        print('Fold:', fold_)
        print(len(trn_idx), len(val_idx))
        tr_x, tr_y = X.iloc[trn_idx, :], y[trn_idx]
        v_X, v_y = X.iloc[val_idx, :], y[val_idx]
        train_data = lgb.Dataset(tr_x, label=tr_y)
        valid_data = lgb.Dataset(v_X, label=v_y)      

        seed_everything(SEED)
        if store_id in ['CA_1', 'CA_2', 'CA_3', 'WI_2']:
            lgb_params['tweedie_variance_power'] = 1.05
        elif store_id in ['TX_1', 'TX_2', 'TX_3', 'WI_3']:
            lgb_params['tweedie_variance_power'] = 1.10
        else:
            lgb_params['tweedie_variance_power'] = 1.20
            
        if store_id in ['CA_2', 'CA_3']:
            lgb_params['n_estimators'] = 1800
        else:
            lgb_params['n_estimators'] = 1400
        
            
        estimator = lgb.train(lgb_params,
                              train_data,
                              valid_sets=[train_data, valid_data],
                              verbose_eval=100)
        # Save model - it's not real '.bin' but a pickle file
        # estimator = lgb.Booster(model_file='model.txt')
        # can only predict with the best iteration (or the saving iteration)
        # pickle.dump gives us more flexibility
        # like estimator.predict(TEST, num_iteration=100)
        # num_iteration - number of iteration want to predict with,
        # NULL or <= 0 means use best iteration
        model_name = 'lgb_model_' + store_id + '_' + str(fold_) + '.bin'
        pickle.dump(estimator, open(model_name, 'wb'))

        # Remove temporary files and objects
        # to free some hdd space and ram memory
        del train_data, valid_data, estimator
        gc.collect()

    # "Keep" models features for predictions
    MODEL_FEATURES = features_columns

########################### Predict
#################################################################################

for fold_ in CV_FOLDS:
    print("FOLD:", fold_)
    # Join back the Test dataset with
    # a small part of the training data
    # to make recursive features
    all_preds = pd.DataFrame()
    base_test = get_base_test()
    # Timer to measure predictions time
    main_time = time.time()

    # Loop over each prediction day
    # As rolling lags are the most timeconsuming
    # we will calculate it for whole day
    for PREDICT_DAY in range(1, 29):
        print('Predict | Day:', PREDICT_DAY)
        start_time = time.time()

        # Make temporary grid to calculate rolling lags
        grid_df = base_test.copy()
        grid_df = pd.concat(
            [grid_df, df_parallelize_run(make_lag_roll, ROLS_SPLIT)], axis=1)

        for store_id in STORES_IDS:

            # Read all our models and make predictions
            # for each day/store pairs
            # model_path = '../data/output/lgb_model_' + store_id + '_' + str(fold_) + '.bin'
            model_path = 'lgb_model_' + store_id + '_' + str(fold_) + '.bin'
            if USE_AUX:
                model_path = AUX_MODELS + model_path

            estimator = pickle.load(open(model_path, 'rb'))

            day_mask = base_test['d'] == (END_TRAIN + PREDICT_DAY)
            store_mask = base_test['store_id'] == store_id

            mask = (day_mask) & (store_mask)
            base_test[TARGET][mask] = estimator.predict(
                grid_df[mask][MODEL_FEATURES])

        # Make good column naming and add
        # to all_preds DataFrame
        temp_df = base_test[day_mask][['id', TARGET]]
        temp_df.columns = ['id', 'F' + str(PREDICT_DAY)]
        if 'id' in list(all_preds):
            all_preds = all_preds.merge(temp_df, on=['id'], how='left')
        else:
            all_preds = temp_df.copy()
        all_preds = all_preds.reset_index(drop=True)
        print('#' * 10,
              ' %0.2f min round |' % ((time.time() - start_time) / 60),
              ' %0.2f min total |' % ((time.time() - main_time) / 60),
              ' %0.2f day sales |' % (temp_df['F' + str(PREDICT_DAY)].sum()))
    all_preds.to_csv('all_preds_' + str(fold_) + '.csv', index=False)
    del temp_df, all_preds

# all_preds_0 = pd.read_csv(AUX_MODELS + 'all_preds_0' + '.csv')
# all_preds_1 = pd.read_csv(AUX_MODELS + 'all_preds_1' + '.csv')
# all_preds_2 = pd.read_csv(AUX_MODELS + 'all_preds_2' + '.csv')


all_preds_0 = pd.read_csv('all_preds_0' + '.csv')
all_preds_1 = pd.read_csv('all_preds_1' + '.csv')
all_preds_2 = pd.read_csv('all_preds_2' + '.csv')

# Create Dummy DataFrame to store predictions
final_all_preds = pd.DataFrame()
final_all_preds['id'] = all_preds_1['id']
for item in all_preds_1:
    if item != 'id':
        final_all_preds[item] = (all_preds_0[item] *
                                 (1 / 3)) + (all_preds_1[item] *
                                             (1 / 3)) + (all_preds_2[item] *
                                                         (1 / 3))
final_all_preds

########################### Export
#################################################################################
# Reading competition sample submission and
# merging our predictions
# As we have predictions only for "_evaluation" data
# we need to do fillna() for "_evaluation" items
submission = pd.read_csv(ORIGINAL + 'sample_submission.csv')[['id']]
submission = submission.merge(final_all_preds, on=['id'], how='left').fillna(0)
submission.to_csv('submission_final' + '.csv', index=False)

# Summary

# Of course here is no magic at all.
# No "Novel" features and no brilliant ideas.
# We just carefully joined all
# our previous fe work and created a model.

# Also!
# In my opinion this strategy is a "dead end".
# Overfits a lot LB and with 1 final submission
# you have no option to risk.

# Improvement should come from:
# Loss function
# Data representation
# Stable CV
# Good features reduction strategy
# Predictions stabilization with NN
# Trend prediction
# Real zero sales detection/classification

# Good kernels references
## (the order is random and the list is not complete):
# https://www.kaggle.com/ragnar123/simple-lgbm-groupkfold-cv
# https://www.kaggle.com/jpmiller/grouping-items-by-stockout-pattern
# https://www.kaggle.com/headsortails/back-to-predict-the-future-interactive-m5-eda
# https://www.kaggle.com/sibmike/m5-out-of-stock-feature
# https://www.kaggle.com/mayer79/m5-forecast-attack-of-the-data-table
# https://www.kaggle.com/yassinealouini/seq2seq
# https://www.kaggle.com/kailex/m5-forecaster-v2
# https://www.kaggle.com/aerdem4/m5-lofo-importance-on-gpu-via-rapids-xgboost

# Features were created in these kernels:
##
# Mean encodings and PCA options
# https://www.kaggle.com/kyakovlev/m5-custom-features
##
# Lags and rolling lags
# https://www.kaggle.com/kyakovlev/m5-lags-features
##
# Base Grid and base features (calendar/price/etc)
# https://www.kaggle.com/kyakovlev/m5-simple-fe

# Personal request
# Please don't upvote any ensemble and copypaste kernels
## The worst case is ensemble without any analyse.
## The best choice - just ignore it.
## I would like to see more kernels with interesting and original approaches.
## Don't feed copypasters with upvotes.

## It doesn't mean that you should not fork and improve others kernels
## but I would like to see params and code tuning based on some CV and analyse
## and not only on LB probing.
## Small changes could be shared in comments and authors can improve their kernel.

## Feel free to criticize this kernel as my knowlege is very limited
## and I can be wrong in code and descriptions.
## Thank you.

final_all_preds.shape
